Event list controller implemented

+++ src/Model/EventListEntry.java
@@ -3,7 +3,8 @@ package Model;
 import java.util.List;
 
 public class EventListEntry {
-    private int t;
+    private double t;
+
     public static int getNextEvent(List<EventListEntry> event){
         return 0;
     }
